Use TypeScript Language for Interactive Window
**Bug**
We currently register a Node.js language service. This causes a `Text Editor -> Node.js` section to be automatically created in the options panel. These builtin pages don't really do anything and users should use the TypeScript options instead. VS15 makes this more clear by renaming the "Typescript" Text editor section to "JavaScript / Typescript"

**fix**
Use the Typescript language service for the interactive window. This is the only place we still use the old Node language service. This also gets us syntax highlighting back for free.

In VS15, delete the node language service entirely. In VS14, we still use some of the submenus for this language service, so we have to keep it around for now.