Extract Project Template Strings for Localization
**Bug**
Strings for project template name and descriptions are currently embedded in the templates files themselves. This prevents them from being localized.

**fix**
Extract all these strings to a resx file to that they can be localized in the future.

**Testing**
Confirmed that all tempaltes still show proper names and descriptions in VS UI.