Remove show all files setting from project tempaltes (#1348)
**Bug**
The `ProjectView` setting is saved inside the `nsjproj` files in our tempaltes. This means that when user alter the soltuion explorer view, the `njsproj` is updated. Since this file may be managed by git, this behavior can be annoying.

**Fix**
Remove the `ProjectView` emlement from the template. If a user turns on 'show all items' now, the setting is saved to a `njsproj.user` file instead.

Closes #1344, Closes #250