Change OnPropertyChangeStrings to use nameof (#1374)
**Bug**
Most of our calls to `OnPropertyChange` currently use string values. This is problematic because strings are not strongly typed, so they can easily get out of sync with the actual property names.

**Fix**
Change these calls to use `nameof` instead. This ensures that an error is produced if the names ever get out of sync. Also this makes c# refactorings do the correct thing.