Move Some Nodejs Strings to Use Resources Directly
**Bug**
We currently go through a `SR` class for accessing localizable strings. This is a pain because when you add a new string, you also have to update the `SR` class. Additionally, it is very easy for the string keys to get out of sync with the resource file.

**Fix**
For the main Node.js project, change basic string use cases to use the `Resources` file directly instead