Use Generated Resource Helper for NodeJs Strings Used in Attributes (#1384)
**Bug**
Follow up on #1382. That PR changed most strings used in code to use the generated resource helper classes. It did not convert over strings that are used in attributes.

**Fix**
Move these strings over as well. The strings passed to attributes must be constant, so `nameof` is used to ensure the strings are strongly typed.