Fix WebLinter Throwing Exception When Opening a Css/Js File inside a Folder
**Bug**
When we open a css/js file inside of a folder in an ntvs project, the webtools linter throws an exception.

I believe the root cause is that the linter calls into `OAProjectItem.Collection` (the implementation of `EnvDTE.ProjectItem.Collection`), we throw a Unsupported Exception. WebTools should handle this (and they took a fix into VS15) but we can also fix this on our side.

**Fix**
Instead of throwing an exception, return null. Other overrides of this property also return null, so I believe this is safe.

**Testing**
Tested originally failing scenario. No exception seen. Also tried opening a bunch of other files and did not see any odd behavior.

Closes #1185