Don't Register as Language(VsTemplate)="JavaScript" in Vs15
**Bug**
The Uwp team contacted us about some of their tests breaking when NTVS is installed in VS15. The root cause seems to be that they attempt to retrieve their "JavaScript" file templates using the vssdk, and this only ends up returning the NTVS templates.

**Fix**
Only in VS15, don't register as a "JavaScript" language template provider.