Fix EnvironmentSetup Script for VS15
**Bug**
The current `EnvironmentSetup.ps1` script does not work foir VS15 Post preview 3. It also does not support multiple installs of VS in different locations (something that is a lot more common in VS15)

**Fix**
Add an optional `vsroot` param that points to where the root dir where VS is installed. Update scripts to use this.