Fix TestAdapter References for Building on VS15
**Bug**
When building NTVS on a machine that only has VS15 post preview 3 installed, `DevEnvDir` is incorrect. The value is supposed to come from the registry, but the registry key does not exist in this case.

**Fix**
Optionally pass DevEnvDir in the buildscript itself insead of looking it up from the registry.