Add test window location to assembly resolver
The adapters are loaded in an out-of-process component for test run. The host process may not
be in the same location as Test Window (e.g. in the newer test platform, this ships in `Common7\IDE\Extensions\TestPlatform`
since it is built as an external component). Adding test window location explicitly so that `Microsoft.VisualStudio.TestWindow.Interfaces` assembly
can be resolved appropriately.