Add support for using javascript unit tests in other project types
This enables the javascript unit test components to work with other
project types than njsproj.

If a project has a TestRoot and a TestFramework specified, we use those
to locate the files containing the unit tests. This also includes changes
to the node project system to make those 2 properties accessible.